Public API Reference

Modeling

Dieter.build_model! — Method

Build the JuMP model describing the optimization problem, specifying the solver to use. The Timestep parameter should be 1 for half-hourly steps, 2 for hourly steps, and 4 for 2-hourly steps. The data must match the time-steps of the Timestep parameter.

Dieter.InitialiseDieterModel — Method

Default constructor for a Dieter model, initialised with the Model DataType and a user-specified data Dict associated with the model in the constructor.

Dieter.initialise_data_file_dict! — Method

Initialise the strings giving the file-path references to the model data. The argument sfx is the file suffix of each model data file. Currently, sfx can be either "csv" or "sql"

Dieter.check_files_exist — Method

Given a Dict of keys => filenames, check at files exist

Dieter.create_relation — Method

Using a DataFrame, create_relation returns a function that can be used to determine a 2-valued relation, that is, a set-to-set correspondence. DataFrame column-names (Symbols) are given that define the set tuples, and a third Indicator column-name is used to indicate set membership. The return type is a function that can be used with two "sets" A and B in tuple2_filter(func,A,B) = filter(x->func(x[1],x[2]), [(a,b) for a in A for b in B])

Dieter.parse_data_to_model! — Method

Build the data for the model, returning intermediate DataFrames containing parsed data. The function will not parse data where the corresponding model setting equals missing.